Refactor parselines utils to suppport context
