+++ src/Data/OrgMode/Parse/Attoparsec/Util/ParseLinesTill.hs
@@ -1,0 +1,112 @@
+{-|
+Module      :  Data.OrgMode.Parse.Attoparsec.Util.ParseLinesTill
+Copyright   :  © 2017 Parnell Springmeyer
+License     :  All Rights Reserved
+Maintainer  :  Parnell Springmeyer <parnell@digitalmentat.com>
+Stability   :  stable
+
+Attoparsec utilities 
+-}
+
+{-# LANGUAGE ScopedTypeVariables #-}
+
+module Data.OrgMode.Parse.Attoparsec.Util.ParseLinesTill (
+  takeALine,
+  ParseLinesTill (..),
+  takeBlockBreak,
+  skipEmptyLines,
+  ) where
+
+import qualified Control.Monad
+import           Control.Arrow         ((&&&))
+import           Data.Semigroup
+import qualified Data.Attoparsec.Text  as Attoparsec.Text
+import           Data.Attoparsec.Text  (Parser, takeTill, isEndOfLine, many1, anyChar, endOfLine, char, isHorizontalSpace, atEnd, parseOnly)
+import           Data.Text             (Text, snoc)
+import qualified Data.Text             as Text
+import           Data.Functor          (($>))
+import           Data.Foldable         (Foldable(..))
+
+takeALine :: Parser Text
+takeALine = do
+  content <- takeTill isEndOfLine
+  Attoparsec.Text.option content (snoc content <$> anyChar)
+
+-- | Match only if there are more content to be consumed.
+--
+-- Opposite of @Attoparsec.Text.endOfInput@
+hasMoreInput :: Parser ()
+hasMoreInput = do
+  x <- atEnd
+  Control.Monad.when x $ fail "reach the end of input"
+
+-- | Matches only if the incoming text line consists nothing or only spaces
+--
+-- A empty line always ends a SectionBlock
+takeEmptyLine :: Parser Text
+takeEmptyLine = Attoparsec.Text.takeWhile isHorizontalSpace <* endOfLine
+
+
+
+-- | Succeed if it is a headline
+headline :: Parser ()
+headline = hasHeadlinePrefix *> atLeastOneSpace where
+  atLeastOneSpace :: Parser ()
+  atLeastOneSpace = do
+    z <- anyChar
+    if isHorizontalSpace z
+      then return ()
+      else fail "Requiring a space after * for headline"
+  hasHeadlinePrefix = many1 (char '*')
+
+-- | Is the current line a @SectionBlock@ break.  A Line is a break
+takeBlockBreak ::  Parser ()
+takeBlockBreak = breakByEmptyLine <> breakByHeadline where
+  breakByEmptyLine = takeEmptyLine $> ()
+  breakByHeadline = headline
+
+-- | Transform a text content as block to work with current parser state
+feedParserText :: Parser s -> Text -> Parser s
+feedParserText  p t =
+  case parseOnly p t of
+    Left s -> fail s
+    Right s -> return s
+
+type Recursive m b a = b -> (b, Parser (m a))
+
+class (Foldable m) => ParseLinesTill m where
+  -- | Fail and reset position when a breaker is found
+  stop :: forall a. Parser (m a) -> Parser (Either () [a])
+  stop p' = do
+    z <- (Right <$> p') <> (return . Left) ()
+    case z of
+      Left _ -> return (Left ())
+      Right x -> if null x
+                  then fail ""
+                  else return . Right . toList $ x
+
+  takeContent :: forall a b. Recursive m b a -> b -> Parser (Text, [a])
+  takeContent next c = tContent <> return (Text.empty, []) where
+    (c', p) = next c
+    tContent = do
+      z <- stop p
+      case z of
+        Right as -> return (Text.empty, as)
+        Left _ -> (\l -> (Text.append l . fst) &&& snd) <$>  takeALine <*> takeContent next c'
+
+  -- | Save the content and parse as the default Plain Text or default Section Paragraph
+  -- and try to parse the new block if the new block exists under the same node
+  parseLinesContextuallyTill :: forall a b. Parser a -> Recursive m b a -> b -> Parser [a]
+  parseLinesContextuallyTill pD next c= skipEmptyLines *> hasMoreInput *> do
+    (content, blocks ) <- takeContent next c
+    if Text.null content && null blocks
+      then fail ""
+      else (: blocks) <$> feedParserText pD content
+
+  parseLinesTill :: forall a. Parser a -> Parser (m a) -> Parser [a]
+  parseLinesTill pDefault pBreaker = parseLinesContextuallyTill pDefault  (const (0 :: Integer , pBreaker)) 0
+
+instance ParseLinesTill (Either a)
+
+skipEmptyLines :: Parser ()
+skipEmptyLines = Attoparsec.Text.many' takeEmptyLine $> ()
